+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Make `/api/v1/timelines/public` return public timeline if `local` is set to `false`.
+
+### Deprecated
+
+- Loading local timeline without `local` parameter.
+
 ## [1.24.1] - 2023-05-15
 
 ### Added